+++ src/MyRouter.jsx
@@ -3,7 +3,8 @@ import { BrowserRouter, Routes, Route } from "react-router-dom";
 import CircularLoader from "./components/loader/CircularLoader";
 
 const MainPage = lazy(() => import("./pages/main/MainPage"));
-const SigninPage = lazy(() => import("./pages/signin/SigninPage"));
+const SigninPage = lazy(() => import("./pages/auth/signin/SigninPage"));
+const SignupPage = lazy(() => import("./pages/auth/signup/SignupPage"));
 
 const MyRouter = () => {
   return (
@@ -12,6 +13,7 @@ const MyRouter = () => {
         <Routes>
           <Route path="/" element={<MainPage />} />
           <Route path="/login" element={<SigninPage />} />
+          <Route path="/signup" element={<SignupPage />} />
         </Routes>
       </BrowserRouter>
     </Suspense>
