--- src/components/signin/SigninHeader.jsx
@@ -1,18 +1,0 @@
-import React from "react";
-import { Avatar, Typography } from "@mui/material";
-import LockOutlinedIcon from "@mui/icons-material/LockOutlined";
-
-const SigninHeader = () => {
-  return (
-    <React.Fragment>
-      <Avatar sx={{ m: 1, bgcolor: "secondary.main" }}>
-        <LockOutlinedIcon />
-      </Avatar>
-      <Typography component="h1" variant="h5">
-        Sign in
-      </Typography>
-    </React.Fragment>
-  );
-};
-
-export default SigninHeader;
